Imporove maintainability
1. getSize now is implemented through hash not switch statement.
   More concise and appear to be much faster.

2. Created helpers.js file, and functions that logically unrelated
   to decoding/encoding will go there.

3. Where possible prefer flat code that nested branches

4. In decodeExt dramatically simplify all the things.

Changes to be committed:
	modified:   lib/decoder.js
	new file:   lib/helpers.js
