+++ lib/decoder.js
@@ -2,14 +2,33 @@
 
 var bl = require('bl')
 var util = require('util')
-
-function IncompleteBufferError (message) {
-  Error.call(this) // super constructor
-  if (Error.captureStackTrace) {
-    Error.captureStackTrace(this, this.constructor) // super helper method to include stack trace in error object
-  }
-  this.name = this.constructor.name
-  this.message = message || 'unable to decode'
+var { IncompleteBufferError } = require('./helpers.js')
+const SIZES = {
+  0xc4: 2,
+  0xc5: 3,
+  0xc6: 5,
+  0xc7: 3,
+  0xc8: 4,
+  0xc9: 6,
+  0xca: 5,
+  0xcb: 9,
+  0xcc: 2,
+  0xcd: 3,
+  0xce: 5,
+  0xcf: 9,
+  0xd0: 2,
+  0xd1: 3,
+  0xd2: 5,
+  0xd3: 9,
+  0xd4: 3,
+  0xd5: 4,
+  0xd6: 6,
+  0xd7: 10,
+  0xd8: 18,
+  0xd9: 2,
+  0xda: 3,
+  0xdb: 5,
+  0xde: 3
 }
 
 util.inherits(IncompleteBufferError, Error)
@@ -18,60 +37,7 @@ module.exports = function buildDecode (decodingTypes) {
   return decode
 
   function getSize (first) {
-    switch (first) {
-      case 0xc4:
-        return 2
-      case 0xc5:
-        return 3
-      case 0xc6:
-        return 5
-      case 0xc7:
-        return 3
-      case 0xc8:
-        return 4
-      case 0xc9:
-        return 6
-      case 0xca:
-        return 5
-      case 0xcb:
-        return 9
-      case 0xcc:
-        return 2
-      case 0xcd:
-        return 3
-      case 0xce:
-        return 5
-      case 0xcf:
-        return 9
-      case 0xd0:
-        return 2
-      case 0xd1:
-        return 3
-      case 0xd2:
-        return 5
-      case 0xd3:
-        return 9
-      case 0xd4:
-        return 3
-      case 0xd5:
-        return 4
-      case 0xd6:
-        return 6
-      case 0xd7:
-        return 10
-      case 0xd8:
-        return 18
-      case 0xd9:
-        return 2
-      case 0xda:
-        return 3
-      case 0xdb:
-        return 5
-      case 0xde:
-        return 3
-      default:
-        return -1
-    }
+    return SIZES[first] || -1
   }
 
   function isValidDataSize (dataLength, bufLength, headerLength) {
@@ -79,10 +45,7 @@ module.exports = function buildDecode (decodingTypes) {
   }
 
   function buildDecodeResult (value, bytesConsumed) {
-    return {
-      value: value,
-      bytesConsumed: bytesConsumed
-    }
+    return { value, bytesConsumed }
   }
 
   function decode (buf) {
@@ -99,8 +62,7 @@ module.exports = function buildDecode (decodingTypes) {
     return result.value
   }
 
-  function tryDecode (buf, offset) {
-    offset = offset === undefined ? 0 : offset
+  function tryDecode (buf, offset = 0) {
     var bufLength = buf.length - offset
     if (bufLength <= 0) {
       return null
@@ -284,12 +246,10 @@ module.exports = function buildDecode (decodingTypes) {
     } else if ((first & 0xe0) === 0xa0) {
       // fixstr up to 31 bytes
       length = first & 0x1f
-      if (isValidDataSize(length, bufLength, 1)) {
-        result = buf.toString('utf8', offset + 1, offset + length + 1)
-        return buildDecodeResult(result, length + 1)
-      } else {
-        return null
-      }
+      if (!isValidDataSize(length, bufLength, 1)) return null
+
+      result = buf.toString('utf8', offset + 1, offset + length + 1)
+      return buildDecodeResult(result, length + 1)
     } else if (first >= 0xe0) {
       // 5 bits negative ints
       result = first - 0x100
@@ -336,13 +296,11 @@ module.exports = function buildDecode (decodingTypes) {
     return buildDecodeResult(result, headerLength + totalBytesConsumed)
   }
 
-  function decodeMap (buf, offset, length, headerLength) {
-    const initialOffset = offset
+  function decodeMap (buf, initialOffset, length, headerLength) {
+    var offset = initialOffset + headerLength
     var result = {}
     var i = 0
-    // var totalBytesConsumed = 0
 
-    offset += headerLength
     while (i++ < length) {
       var keyResult = tryDecode(buf, offset)
       if (!keyResult) return null
@@ -353,11 +311,7 @@ module.exports = function buildDecode (decodingTypes) {
       offset += valueResult.bytesConsumed
 
       result[keyResult.value] = valueResult.value
-      // totalBytesConsumed += keyResult.bytesConsumed + valueResult.bytesConsumed
     }
-    
-    // Well, :)
-    // console.log(headerLength + totalBytesConsumed === offset - initialOffset)
 
     return buildDecodeResult(result, offset - initialOffset)
   }
@@ -394,30 +348,20 @@ module.exports = function buildDecode (decodingTypes) {
     return buildDecodeResult(new Date(millis), size + headerSize)
   }
 
-  function decodeExt (buf, offset, type, size, headerSize) {
-    var i,
-      toDecode
-
-    offset += headerSize
+  function decodeExt (buf, initialOffset, type, size, headerSize) {
+    const offset = initialOffset + headerSize
+    const toDecode = buf.slice(offset, offset + size)
 
     // Pre-defined
-    if (type < 0) { // Reserved for future extensions
-      switch (type) {
-        case -1: // Tiemstamp https://github.com/msgpack/msgpack/blob/master/spec.md#timestamp-extension-type
-          toDecode = buf.slice(offset, offset + size)
-          return decodeTimestamp(toDecode, size, headerSize)
-      }
-    }
+    if (type === -1) return decodeTimestamp(toDecode, size, headerSize)
+    // Reserved for future extensions
+    if (type < 0) throw new Error('unable to find ext type ' + type)
 
-    for (i = 0; i < decodingTypes.length; i++) {
-      if (type === decodingTypes[i].type) {
-        toDecode = buf.slice(offset, offset + size)
-        var value = decodingTypes[i].decode(toDecode)
-        return buildDecodeResult(value, headerSize + size)
-      }
-    }
+    const codec = decodingTypes.find(it => it.type === type)
+    if (!codec) throw new Error('unable to find ext type ' + type)
+    var value = codec.decode(toDecode)
 
-    throw new Error('unable to find ext type ' + type)
+    return buildDecodeResult(value, headerSize + size)
   }
 }
